Auth - new service model auth - Repo.currentProfile is now Auth.instance.profile - fix 'uid called on null' and related errors stemming from Repo.currentProfile

+++ lib/utils/image_file_bundle.dart
@@ -43,7 +43,7 @@ class ImageUrlBundle {
       small: map['small'] ?? '',
       medium: map['medium'] ?? '',
       original: map['original'] ?? '',
-      aspectRatio: map['aspect_ratio'] ?? 1,
+      aspectRatio: (map['aspect_ratio'] ?? 1.0).toDouble(),
     );
   }
 }
